# backend/src/main/java/com/expensesplit/dto/UpdateGroupRequest.java
package com.expensesplit.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UpdateGroupRequest {
    @NotBlank(message = "Group name cannot be blank")
    private String name;
}